server/pyproject.toml
[project]
name = "riglet"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2",
    "pyyaml",
    "numpy",
    "pyserial",
    "sse-starlette",
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "types-pyyaml>=6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
